Improved algorithm execution speed

See merge request datascience/valveinsight!1

+++ algorithms_c/ValveInsightPlatformIO/lib/switchingDetection/src/switchingDetectionFixed.h
@@ -5,13 +5,13 @@ extern "C" {
 #endif
 
 #include <stdint.h>
+#include <stdbool.h>
 
 /** @brief Reset internal state and clear ADC buffer. */
 void switchingDetectionFixed_Reset(void);
 
 void switchingDetectionFixed_StoreADC(int16_t adcValue);
-
-double switchingDetectionFixed_Calculate(double timestep);
+float switchingDetectionFixed_Calculate(float timestep);
 
 #ifdef __cplusplus
 }
